Integrate changes from master (#36)
* Feat: Add deep linking - syrius://

* Update main_app_container.dart

* fix: MacOS deep link support

* Feat: Add UI for Wallet Connect

* feat: Enable pairing with dApps through WC

* chore: Add namespaces and more listeners

* Replace bodyText1 with bodyLarge

* fix: Change WC initialization

* chore: Update gitignore

* chore: Update repo after gitignore update

* feat: Add method that activates the pairing

* feat: Add feature that captures a part of the screen and listens to the clipboard

* feat: Build table that shows the current pairings

* feat: Add feature to scan the QR code from the screen

* fix: Remove Expanded widget from Pairings card widget tree

* feat: Listens to values in the clipboard in order to check for WC URI

* fix: Make access check for MacOS before scanning screen

* feat: Redirect to WC tab when a WC URI is detected

* fix: Rebuild WC URI input field when another URI is detected in the clipboard

* Updated native libraries

* fix: Recreate the windows folder

* feat: Add code that opens the app through deep linking

* refactor: Change API call when activating a pairing

* feat: Add possibility to deactivate a WC pair

* feat: Redirect to WC tab when opening the wallet through WC deep linking

* fix: Error initializing the WC uri input field

* feat: Improve validating the WC URI

* feat: Add handler for znn_net_info WC method

* feat: Add handler for znn_sign WC method

* refactor: Create function for signing with the key pair from the wallet file

* feat: Add handler for WC znn_send method

* feat: Change name of znn_net_info method to znn_info and refactor sending a block

* fix: Restyle the Topic column from the WC pairings table

* feat: Send back an account block to the dApp after executing the znn_send method

* feat: Send events to the dApp when the address or chain id were changed

* feat: Send error to the dApp when user rejects the transfer

* fix: Change the error type send to the dApp

* feat: Add confirmation dialogs before making an action sent from the dApp

* refactor

* feat: Fix dialog when dApp wants to send a tx

* fix: Reset the URI value from the change notifier once it's used

* fix: Don't redirect to WC when detecting a WC URI if the wallet is locked

* fix: Show WC URI detected notification if the user is already on the WC tab

* feat: Throw an error when trying to trigger an action through WC and wallet is locked

* feat: Reconnect to websocket when WC connection falls

* feat: Add methods to retrieve the active sessions and all sessions

* feat: Get the reference for getting the ZNN SDK

* feat: Change position of the WC Connect button and disable it in case the input is not valid

* refactor: Delete unused method

* feat: Send notification when user cancels the QR scanning

* feat: Keep the WC URI in memory when wallet is locked

* feat: Update WalletConnectPairingCard with the needed permissions

* feat: Update files

* feat: Wrap up WC

* fix: Clean up

* feat: Get WC project id from environment variable

* fix: Rebuild 'Connect' button when the URI input field is also rebuild

* feat: Rebuild pubspec.lock

* feat: Change the source of the ZNN SDK

* build: referenced wallet-connect sdk branch

* build: added required macos entitlements

* refactor: small refactor

* docs: updated README

* cicd: retrieve WALLET_CONNECT_PROJECT_ID

* cicd: added environment name

* cicd: updated flutter-action

* cicd: set linux directory permissions

* feat: added BigInt support

* fix: referenced latest SDK changes

* fix: desktop window

* fix: windows desktop window

* fix: bigInt decimal conversions

* chore: updated dependencies

* fix: tokenSymbol widget

* fix: minor UI fixes

* cicd: hardcoded flutter version for compatibility

* feat: Change caching the active sessions for a pairing

* fix: allow any ZTS in znn_send

* feat: added WalletConnect events

* fix: remove relayClient reconnect

* feat: Add widget for pairing session data

* feat: Tweak WC pairing flow

* fix: Emit the session event (address and chain id events) to each dApp with an active pair and session

* fix: add windows directory

* chore: update walletconnect_flutter_v2 to v2.0.4

* fix: BigInt fixes

* feat: Regenerate windows folder

* feat: Regenerate windows folder

* feat: Refresh WC URI input field after pairing

* feat: error notification for auto-receive

* chore: updated walletconnect_flutter_v2 to v2.0.6

* feat: Build WC uri card

* feat: Build WC QR card

* feat: Build WC camera card

* feat: Put the pairings and sessions widgets side by side

* feat: Change button name

* fix: flutter windows directory

* chore: update uni_links_desktop to v0.1.6

* fix: windows deep link handling

* feat: added notification for incoming links

* feat: Refresh pairing and session lists when connecting to dApp

* feat: Align pairing and session list widgets on top of each other

* refactor: Add the getter pairings to the WalletConnectService class

* feat: replaced uni_links with app_links

* fix: Revert changes to the pairing list widget

* fix: Make InfiniteScrollBloc handle case when the data request is not paginated

* fix: Don't dispose the pairings and sessions bloc

* fix: Also refresh sessions list after deleting a pairing

* fix: Check for broken WC URI

* fix: added refresh for onPairingCreate

* fix: Redirect to WC tab when an WC URI is detected while the wallet is locked

* fix: Don't cache results for blocs that don't handle paginated results

* fix: Decode WC URI before parsing

* feat: Upgrade project to Dart 3.0.1

* feat: Delete button that disconnects the WC session with the dApp

* feat: Throw error if trying to send tx from dApp while wallet is locked

* cicd: upgraded flutter to v3.10.2

* chore: update texts

* feat: Update WC dependency to 2.0.8

* fix: parse WalletConnect URI

* fix: Dart lint warnings

* feat: bump version to v0.0.7

* fix: Windows deep link

* Updated native libraries

* feat: Hard code chainId to 1 for WC

* fix: WalletConnect improvements

* fix: When to mark last page of the infinite scroll bloc

* fix: dialog navigator pop

* fix: added marquee for token balance

* feat: logging to disk

* refactor: renamed votePillar to delegateToPillar

* fix: WalletConnect issues

* feat: programmatically display WalletConnect tab

* feat: added deep link actions

* feat: update defaultValue for kWcProjectId

* chore: updated walletconnect_flutter_v2 to v2.0.12

* chore: update dependencies

* feat: added camera QR scanning for MacOS

* fix: WakelockPlus Linux support

* fix: replaced logger with Logger

* chore: added descriptions and code cleanup

* fix: wallet_connect_tab_child  import

* fix: minimum amount to send

* fix: refresh WalletConnectSessionsBloc

* fix: remainingZnnBudget decimals

* fix: onSessionDelete refresh

* Integrate changes from master with P2P swaps

---------

+++ lib/widgets/modular_widgets/p2p_swap_widgets/p2p_swaps_list_item.dart
@@ -135,7 +135,7 @@ class _P2pSwapsListItemState extends State<P2pSwapsListItem> {
   }
 
   Widget _getAmountWidget(
-      int? amount, int? decimals, String? tokenStandard, String? symbol) {
+      BigInt? amount, int? decimals, String? tokenStandard, String? symbol) {
     if (amount == null ||
         decimals == null ||
         tokenStandard == null ||
